+++ fuse4dbricks/main.py
@@ -18,7 +18,7 @@ from fuse4dbricks.fs.inode_manager import InodeManager
 from fuse4dbricks.fs.metadata_manager import MetadataManager
 from fuse4dbricks.fs.auth_manager import AuthManager
 from fuse4dbricks.fs.operations import UnityCatalogFS
-from fuse4dbricks.auth.provider import AuthProvider
+from fuse4dbricks.auth.provider import AuthProvider, PositWorkbenchAuthProvider
 from fuse4dbricks.auth.entra_id import EntraIDPublicAuthProvider
 from fuse4dbricks.storage.persistence import DiskPersistence, clear_cache
 
@@ -48,10 +48,6 @@ def parse_args():
     parser.add_argument(
         "--workspace", default="", help="https://adb-xxxx.azuredatabricks.net"
     )
-    parser.add_argument("--tenant-id", help="Azure Tenant ID (required for device auth)", required=False, default="")
-    parser.add_argument(
-        "--client-id", default="", required=False, help="Azure App Client ID (required for device auth)"
-    )
     parser.add_argument("mountpoint", help="Local directory to mount")
     parser.add_argument(
         "--disk-cache-dir", default=_get_default_cache_dir(), help="Local disk cache location"
@@ -60,7 +56,9 @@ def parse_args():
     parser.add_argument("--allow-other", action="store_true",
         help="Use -o allow_other only if /etc/fuse.conf has user_allow_other"
     )
-
+    parser.add_argument("--posit-workbench", action="store_true", default=False,
+        help="Use the Posit Workbench databricks integration to obtain a token"
+    )
     parser.add_argument(
         "--clear-cache", action="store_true", help="Clear disk cache on startup"
     )
@@ -173,9 +171,8 @@ async def async_main():
     # Auth
     logging.info("Initializing Authentication...")
     external_provider = None
-    if args.tenant_id and args.client_id:
-        external_provider = EntraIDPublicAuthProvider(tenant_id=args.tenant_id, client_id = args.client_id, cache_dir=auth_cache_dir)
-
+    if args.positworkbench:
+        external_provider = PositWorkbenchAuthProvider()
     auth_provider: AuthProvider = AuthProvider(provider=external_provider)
     # Init Components
     uc_client = UnityCatalogClient(workspace, auth_provider)